Create a simulation unit (USM) for the STICS model

Uses a pre-existing USM, copy files to target folder(s) of simulation, set the USMs ready for simulation. import_usm create as many sub-folders as USM neede in the target folder.

import_usm(dir.orig = NULL, dir.targ = getwd(), stics = NULL,
  usm_name = NULL, overwrite = T, verbose = NULL)

Arguments

dir.orig
Path to the directory from which copy the simulation files. If NULL (the default), uses the package dummy USM.

dir.targ
Path to the target directory for simulation. Created if missing.

stics
Path to the STICS model executable (optional, only needed if not present in dir.orig)

usm_name
Vector name of the USM(s).

overwrite
Boolean. Overwrite files and folders if already present. See details.

verbose
Boolean. Does the function output success and failure messages ?

Details

This function is a helper function used by other package functions. If overwrite= F, the function show to the user which files are already present, and asks the user what to do, so be careful while using this function for programming.

Examples

# NOT RUN {
# Set a dummy simulation in the folder "1-Simulations", relative to the
# project path:

library(sticRs)
import_usm(dir.targ = "1-Simulations")

# }