Internal function to set some STICS input file parameters

Replace or set an input parameter from a pre-existing STICS input file. This function is called by some of the generic set_* functions under the hood.

set_file(filepath, param, value, add)

Arguments

filepath
Path to the parameter file

param
Parameter name

value
New parameter value

add
Boolean. Append input to existing file (add to the list)

Details

The function uses sys.call to know from which function of the set_* family it is called, so it won't work properly if called by the user directly. This is why this function is internal.

Note

This function is not used for set_soil.

See also